Refactored ControllerProfileActivity
Refactored ControllerProfileActivity into a base class with 2 derived
classes.

Also, now using the layout-land resource folder for the portrait version
of this activity.
