+++ res/layout/controller_profile_activity.xml
@@ -30,68 +30,53 @@
         android:gravity="center"
         android:orientation="vertical" >
 
-        <RelativeLayout
-            android:id="@+id/all_pads"
-            android:layout_width="match_parent"
+        <include
+            android:id="@+id/include_b_pad"
+            android:layout_width="@dimen/inputMapPreferenceButtonSizeX5"
+            android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
+            layout="@layout/b_pad" />
+
+        <include
+            android:id="@+id/include_d_pad"
+            android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
+            layout="@layout/d_pad" />
+
+        <include
+            android:id="@+id/include_a_pad"
+            android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
+            layout="@layout/a_pad" />
+
+        <include
+            android:id="@+id/include_c_pad"
+            android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
+            android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
+            layout="@layout/c_pad" />
+
+        <TableRow
+            android:id="@+id/tableRowFaceButtons"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:focusable="false"
-            android:focusableInTouchMode="false" >
-
-            <include
-                android:id="@+id/include_b_pad"
-                android:layout_width="@dimen/inputMapPreferenceButtonSizeX5"
-                android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_alignParentTop="true"
-                android:layout_centerHorizontal="true"
-                layout="@layout/b_pad" />
-
-            <include
-                android:id="@+id/include_d_pad"
-                android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_alignTop="@+id/include_b_pad"
-                android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
-                android:layout_toLeftOf="@+id/include_b_pad"
-                layout="@layout/d_pad"
-                tools:ignore="RtlHardcoded" />
-
-            <include
-                android:id="@+id/include_c_pad"
-                android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_alignTop="@+id/include_b_pad"
-                android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
-                android:layout_toRightOf="@+id/include_b_pad"
-                layout="@layout/c_pad"
-                tools:ignore="RtlHardcoded" />
-
-            <include
-                android:id="@+id/include_a_pad"
-                android:layout_width="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_height="@dimen/inputMapPreferenceButtonSizeX3"
-                android:layout_below="@id/include_b_pad"
-                android:layout_centerHorizontal="true"
-                android:layout_marginTop="@dimen/inputMapPreferenceButtonSize"
-                layout="@layout/a_pad" />
+            android:layout_marginTop="@dimen/inputMapPreferenceButtonSize" >
 
             <Button
                 android:id="@+id/buttonB"
                 style="@style/btnMappable"
-                android:layout_alignBottom="@+id/include_c_pad"
-                android:layout_toLeftOf="@+id/include_c_pad"
                 android:background="@drawable/button_b"
-                android:text="@string/inputMapActivity_btnB"
-                tools:ignore="RtlHardcoded" />
+                android:text="@string/inputMapActivity_btnB" />
 
             <Button
                 android:id="@+id/buttonA"
                 style="@style/btnMappable"
-                android:layout_below="@+id/buttonB"
-                android:layout_toRightOf="@+id/buttonB"
+                android:layout_marginLeft="@dimen/inputMapPreferenceButtonSize"
                 android:background="@drawable/button_a"
                 android:text="@string/inputMapActivity_btnA"
                 tools:ignore="RtlHardcoded" />
-        </RelativeLayout>
+        </TableRow>
 
         <EditText
             android:id="@+id/dummyImeListener"
